insert gdp procedure docs into runtests
gleaned from lib/game-data-packager-shared, the
procedure documentation says what the procedures "should do".
They are useful reference to write tests against, so insert
a copy of them into the "runtests" script. It also indicates
in one file which procedures have no tests written for them.

There is a risk the text will get out of sync with the text
in the actual procedure definitions...